ERROR
FLT_MAX and FLT_MIN not defined
